index changed

+++ admin/patient_list.php
@@ -72,9 +72,9 @@
                                             <a title="Update" href="patient_edit.php?id=<?= $d->id ?>">
                                                 <i class="fa fa-edit"></i>
                                             </a>
-                                            <a title="Delete" class="text-danger" href="patient_delete.php?id=<?= $d->id ?>">
-                                                <i class="fa fa-trash"></i>
-                                            </a>
+                                            <a title="Delete" onclick="return confirm('Are you sure to delete this!')" class="text-danger" href="patient_delete.php?id=<?= $d->id ?>">
+                                            <i class="fa fa-trash"></i>
+                                        </a>
                                         </td>
                                     </tr>
                                 <?php
